feat: local data source implementation and add generate task use case

+++ app/src/main/java/io/github/nuarz71/todolist/android/data/local/TaskDao.kt
@@ -22,7 +22,7 @@ internal interface TaskDao {
     fun getAllToDo(): Flow<List<TaskEntity>>
     
     @Query("DELETE FROM todo WHERE id=:id")
-    fun deleteTaskById(id: Long): Long
+    fun deleteTaskById(id: Long): Int
     
     @Query("DELETE FROM todo")
     fun deleteAllTask()
